+++ CHANGELOG.md
@@ -28,6 +28,7 @@ See [Mixins/collections.py](src/runtime/Mixins/collections.py).
 and other `PyObject` derived types when called from Python.
 -   .NET classes, that have `__call__` method are callable from Python
 -   `PyIterable` type, that wraps any iterable object in Python
+-   `PythonEngine` properties for supported Python versions: `MinSupportedVersion`, `MaxSupportedVersion`, and `IsSupportedVersion`
 
 
 ### Changed